quicktranslate-backend/requirements.txt
fastapi
uvicorn
transformers
python-dotenv
